+++ src/main/kotlin/calculator/ExpressionChecker.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 object ExpressionChecker {
     fun isEmpty(number: List<Int>, expression: List<Operator>) = number.isNullOrEmpty() || expression.isNullOrEmpty()
