+++ src/main/kotlin/calculator/Main.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 fun main() {
     StringCalculator.parsing(
