+++ src/main/kotlin/calculator/TypeExtractor.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 object TypeExtractor {
