+++ src/main/kotlin/person/Person.kt
@@ -1,3 +1,3 @@
-package step1
+package person
 
 class Person(val name: String, val age: Int?, var nickname: String?)
