+++ src/test/kotlin/person/PersonTest.kt
@@ -1,10 +1,9 @@
-package step1
+package person
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
-import step1.Person
 
 class PersonTest {
